+++ src/commonMain/kotlin/TagLogger.kt
@@ -1,0 +1,10 @@
+package dev.inmo.kslog.common
+
+import kotlin.jvm.JvmInline
+
+@JvmInline
+value class TagLogger(val tag: String) : KSLog {
+    override fun performLog(level: LogLevel, tag: String?, message: String, throwable: Throwable?) {
+        KSLog.performLog(level, tag ?: tag, message, throwable)
+    }
+}