+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.2
+
+* New logger `TagLogger` and now extension `Any#logger` will use it to decrease objects allocations
+* Now it is possible to create `KSLog` using any iterable
+* Now it is possible to create `KSLog` using vararg log levels
+
 ## 0.0.1
 
 * Project has been inited